Make optimizer passes configurable. API is with list in config file. Splitting hls_model.py into hls_model.py and hls_layers.py was necessary to remove circular import dependency from optimizers importing Layers and utilities, while hls_model now needs to import optimizer
