+++ hls4ml/converters/pytorch_to_hls.py
@@ -8,7 +8,6 @@ import pickle
 import re
 
 from hls4ml.model import HLSModel
-from hls4ml.model.optimizer import optimize_model
 
 class PyTorchDataReader:
     def __init__(self, config):
@@ -101,6 +100,4 @@ def pytorch_to_hls(yamlConfig):
     reader = PyTorchDataReader(yamlConfig)
     print('Creating HLS model')
     hls_model = HLSModel(yamlConfig, reader, layer_list)
-    optimizers = ['eliminate_linear_activation', 'merge_batch_norm_quantized_tanh', 'quantize_dense_output', 'fuse_dense_batch_norm']
-    optimize_model(hls_model, optimizers)
     return hls_model
